Add files via upload
Hello guys! So our team participated in a hackathon in VVCE Mysore. We chose forest fires as our topic, basically trained an AI model to detect fires in forests through areal imagery. Being new in the field of training a model, we surely took a lot of help from chatGPT 4.0 and github. First, trained the model to identify forest, non-forest images(which was not needed), later fire and non-fire images. We got the datasets from kaggle. Later the model was trained on a video, once fire is captured on a frame in the video it returns that frame with it's serenity level. This can be sent to any device a forest ranger wants. It helps to locate the fire, so time saving. This is the gist of what we wanted to do.